requirements.txt
maafw>=2.2.0b1
